feat: moving `api-core` back into the main `api` package
