+++ packages/api/package.json
@@ -29,15 +29,18 @@
     "node": "^14 || ^16"
   },
   "dependencies": {
-    "@readme/api-core": "file:../core",
     "@readme/oas-to-har": "^16.0.0",
     "@readme/openapi-parser": "^2.0.0",
+    "datauri": "^4.1.0",
     "fetch-har": "^6.1.0",
     "find-cache-dir": "^3.3.1",
     "form-data-encoder": "^1.7.1",
     "formdata-node": "^4.3.2",
+    "get-stream": "^6.0.1",
     "isomorphic-fetch": "^3.0.0",
     "js-yaml": "^4.1.0",
+    "json-schema-traverse": "^1.0.0",
+    "lodash.merge": "^4.6.2",
     "make-dir": "^3.1.0",
     "oas": "^18.0.0"
   },
@@ -47,9 +50,9 @@
     "@types/chai": "^4.3.0",
     "@types/find-cache-dir": "^3.2.1",
     "@types/js-yaml": "^4.0.5",
+    "@types/lodash.merge": "^4.6.6",
     "@types/mocha": "^9.1.0",
     "chai": "^4.3.6",
-    "datauri": "^4.1.0",
     "eslint": "^8.10.0",
     "mocha": "^9.2.1",
     "mock-require": "^3.0.3",
