+++ packages/api/src/core/getJSONSchemaDefaults.ts
