+++ packages/api/src/core/prepareAuth.ts
