+++ packages/api/test/core/getJSONSchemaDefaults.test.ts
@@ -1,5 +1,5 @@
 import { expect } from 'chai';
-import getJSONSchemaDefaults from '../src/lib/getJSONSchemaDefaults';
+import getJSONSchemaDefaults from '../../src/core/getJSONSchemaDefaults';
 import Oas from 'oas';
 
 describe('#getJSONSchemaDefaults()', function () {
