+++ packages/api/test/core/prepareParams.test.ts
@@ -1,9 +1,9 @@
 import { assert, expect } from 'chai';
 import fs from 'fs';
 import Oas from 'oas';
-import prepareParams from '../src/lib/prepareParams';
+import prepareParams from '../../src/core/prepareParams';
 
-import payloadExamples from './__fixtures__/payloads.oas.json';
+import payloadExamples from '../__fixtures__/payloads.oas.json';
 
 describe('#prepareParams', function () {
   let fileUploads: Oas;
@@ -148,9 +148,9 @@ describe('#prepareParams', function () {
     });
 
     describe('image/png', function () {
-      it('should support a relative file path payload', async function () {
+      it('should support a file path payload', async function () {
         const operation = fileUploads.operation('/anything/image-png', 'post');
-        const body = `${__dirname}/__fixtures__/owlbert.png`;
+        const body = `${__dirname}/../__fixtures__/owlbert.png`;
 
         const res = await prepareParams(operation, body);
         expect(res.body).to.contain('data:image/png;name=owlbert.png;base64,');
