--- packages/core/.gitignore
@@ -1,4 +1,0 @@
-.nyc_output/
-@types/
-dist/
-node_modules/
