--- packages/core/.mocharc.json
@@ -1,13 +1,0 @@
-{
-  "require": [
-    "test/helpers/init.js",
-    "ts-node/register"
-  ],
-  "watch-extensions": [
-    "ts"
-  ],
-  "watch-files": ["src/**/*.ts", "test/**/*.ts"],
-  "recursive": true,
-  "reporter": "spec",
-  "timeout": 60000
-}
