--- packages/core/.npmignore
@@ -1,6 +1,0 @@
-.nyc_output/
-test/
-.eslint*
-.gitignore
-.mocharc.json
-.prettier*
