--- packages/core/.prettierignore
@@ -1,3 +1,0 @@
-.nyc_output/
-@types/
-dist/
