--- packages/core/README.md
@@ -1,7 +1,0 @@
-# @readme/api-core
-
-Core functionality behind ReadMe's [api](https://npm.im/api) SDK framework.
-
-[![npm](https://img.shields.io/npm/v/@readme/api-core)](https://npm.im/@readme/api-core) [![Build](https://github.com/readmeio/api/workflows/CI/badge.svg)](https://github.com/readmeio/api)
-
-[![](https://d3vv6lp55qjaqc.cloudfront.net/items/1M3C3j0I0s0j3T362344/Untitled-2.png)](https://readme.io)
