--- packages/core/package.json
@@ -1,58 +1,0 @@
-{
-  "name": "@readme/api-core",
-  "version": "4.2.0",
-  "description": "Core functionality behind ReadMe's API SDK framework",
-  "main": "dist/index.js",
-  "types": "@types/index.d.ts",
-  "scripts": {
-    "build": "tsc",
-    "lint": "eslint . --ext .js,.ts",
-    "prebuild": "rm -rf dist/ @types/",
-    "prepack": "npm run build",
-    "pretest": "npm run lint",
-    "prettier": "prettier --list-different --write \"./**/**.{js,ts}\"",
-    "test": "nyc mocha \"test/**/*.test.ts\""
-  },
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/readmeio/api.git",
-    "directory": "packages/core"
-  },
-  "bugs": {
-    "url": "https://github.com/readmeio/api/issues"
-  },
-  "author": "Jon Ursenbach <jon@readme.io>",
-  "license": "MIT",
-  "engines": {
-    "node": "^14 || ^16"
-  },
-  "dependencies": {
-    "@readme/oas-to-har": "^16.0.0",
-    "datauri": "^4.1.0",
-    "get-stream": "^6.0.0",
-    "isomorphic-fetch": "^3.0.0",
-    "json-schema-traverse": "^1.0.0",
-    "lodash.merge": "^4.6.2",
-    "oas": "^18.0.0"
-  },
-  "devDependencies": {
-    "@readme/eslint-config": "^8.5.0",
-    "@readme/oas-examples": "^4.4.0",
-    "@types/chai": "^4.3.0",
-    "@types/lodash.merge": "^4.6.6",
-    "@types/mocha": "^9.1.0",
-    "chai": "^4.3.6",
-    "eslint": "^8.10.0",
-    "mocha": "^9.2.1",
-    "nyc": "^15.1.0",
-    "prettier": "^2.5.1",
-    "typescript": "^4.5.5"
-  },
-  "prettier": "@readme/eslint-config/prettier",
-  "nyc": {
-    "exclude": [
-      "dist/",
-      "test/"
-    ]
-  }
-}
