--- packages/core/src/index.ts
@@ -1,7 +1,0 @@
-import getJSONSchemaDefaults from './lib/getJSONSchemaDefaults';
-import parseResponse from './lib/parseResponse';
-import prepareAuth from './lib/prepareAuth';
-import prepareParams from './lib/prepareParams';
-import prepareServer from './lib/prepareServer';
-
-export { getJSONSchemaDefaults, parseResponse, prepareAuth, prepareParams, prepareServer };
