--- packages/core/test/__fixtures__/hello.jp.txt
@@ -1,1 +1,0 @@
-速い茶色のキツネは怠惰な犬を飛び越えます
