--- packages/core/test/__fixtures__/hello.txt
@@ -1,1 +1,0 @@
-Hello world!
