--- packages/core/test/tsconfig.json
@@ -1,8 +1,0 @@
-{
-  "extends": "../tsconfig.json",
-  "compilerOptions": {
-    "noImplicitAny": false,
-    "resolveJsonModule": true
-  },
-  "include": ["../src/**/*", "*.ts", "**/*"]
-}
