--- packages/core/tsconfig.json
@@ -1,13 +1,0 @@
-{
-  "compilerOptions": {
-    "allowJs": true,
-    "baseUrl": "./src",
-    "declaration": true,
-    "declarationDir": "./@types",
-    "esModuleInterop": true,
-    "lib": ["dom", "es2020"],
-    "noImplicitAny": true,
-    "outDir": "dist"
-  },
-  "include": ["./src/**/*"]
-}
